Initial Commit
This is a sample Facebook application that uses the new Insights API.  It is
built to run on Google App Engine.  The application fetches the pages and
applications owned by the logged-in user.  The user then selects an object and a
date for which to fetch Insights data.  The application then downloads Insights
data for the specified object, formats it as CSV, and returns this file to the
user.